+++ code-gen/update-cluster-state.sh
@@ -171,7 +171,9 @@ ${DATASYNC_P1AS_SYNC_SERVER}
 ${LEGACY_LOGGING}
 ${ARGOCD_SLACK_TOKEN_BASE64}
 ${RADIUS_PROXY_ENABLED}
-${PF_PROVISIONING_ENABLED}'
+${PF_PROVISIONING_ENABLED}
+${SLACK_CHANNEL}
+${PROM_SLACK_CHANNEL}'
 
 ########################################################################################################################
 # Export some derived environment variables.